Move ProofStore to separate file
