+++ src/merk/mod.rs
@@ -3,6 +3,8 @@ mod client;
 mod ics23;
 #[cfg(feature = "merk-full")]
 mod proofbuilder;
+#[cfg(feature = "merk-verify")]
+pub mod proofstore;
 #[cfg(feature = "merk-full")]
 pub mod snapshot;
 #[cfg(feature = "merk-full")]
@@ -12,6 +14,7 @@ pub use client::Client;
 pub use merk;
 #[cfg(feature = "merk-full")]
 pub use proofbuilder::ProofBuilder;
+#[cfg(feature = "merk-verify")]
+pub use proofstore::ProofStore;
 #[cfg(feature = "merk-full")]
 pub use store::MerkStore;
-pub use store::ProofStore;
